make `opentelemetry` optional dependency
