+++ mindsdb/api/mysql/mysql_proxy/mysql_proxy.py
@@ -83,7 +83,7 @@ from mindsdb.api.mysql.mysql_proxy.utilities.lightwood_dtype import dtype
 from mindsdb.utilities import log
 from mindsdb.utilities.config import config
 from mindsdb.utilities.context import context as ctx
-from mindsdb.utilities.otel.metric_handlers import get_query_request_counter
+from mindsdb.utilities.otel import increment_otel_query_request_counter
 from mindsdb.utilities.wizards import make_ssl_cert
 
 logger = log.getLogger(__name__)
@@ -558,9 +558,7 @@ class MysqlProxy(SocketServer.BaseRequestHandler):
             )
 
         # Increment the counter and include metadata in attributes
-        metadata = ctx.metadata(query=sql)
-        query_request_counter = get_query_request_counter()
-        query_request_counter.add(1, metadata)
+        increment_otel_query_request_counter(ctx.get_metadata(query=sql))
 
         return resp
