+++ mindsdb/utilities/context.py
@@ -54,7 +54,7 @@ class Context:
     def load(self, storage: dict) -> None:
         self._storage.set(storage)
 
-    def metadata(self, **kwargs) -> dict:
+    def get_metadata(self, **kwargs) -> dict:
         return {
             'user_id': self.user_id or "",
             'company_id': self.company_id or "",
