+++ requirements/requirements-opentelemetry.txt
@@ -1,0 +1,6 @@
+opentelemetry-api==1.27.0
+opentelemetry-sdk==1.27.0
+opentelemetry-exporter-otlp==1.27.0
+opentelemetry-instrumentation-requests==0.48b0
+opentelemetry-instrumentation-flask==0.48b0
+opentelemetry-distro==0.48b0
